Added option for multiple email notification addresses.
Added logic change also to opzet script that when debug is true it will 
always perform an Website check making it easier to debug, so do not 
forget to set it back to false when not debugging!
afvalwijzerconfig.lua is not working when the workdirectory is not the 
script directory so removed this option.